Moving retrieve_all_comments and related functions from application store to assessment store
