--- pre_award/application_store/db/queries/comments/__init__.py
@@ -1,1 +1,0 @@
-from pre_award.application_store.db.queries.comments.queries import *  # noqa
