+++ tests/pre_award/application_store_tests/test_submit.py
@@ -1,12 +1,9 @@
-import io
 import json
-from collections import Counter
-from datetime import datetime, timedelta, timezone
+from datetime import datetime, timezone
 from unittest import mock
 from unittest.mock import MagicMock
 from uuid import uuid4
 
-import pandas as pd
 import pytest
 from click.testing import CliRunner
 from flask_session import Session
@@ -24,7 +21,6 @@ from pre_award.application_store.db.queries.application.queries import (
     submit_application,
     update_application_fields,
 )
-from pre_award.application_store.db.queries.comments.queries import export_comments_to_excel, retrieve_all_comments
 from pre_award.application_store.db.queries.form.queries import add_new_forms
 from pre_award.application_store.db.queries.updating.queries import update_form
 from pre_award.application_store.external_services import get_fund
@@ -33,14 +29,12 @@ from pre_award.assessment_store.config.mappings.assessment_mapping_fund_round im
 from pre_award.assessment_store.db.models.assessment_record.assessment_records import AssessmentRecord
 from pre_award.assessment_store.db.models.assessment_record.enums import Status
 from pre_award.assessment_store.db.models.assessment_record.enums import Status as WorkflowStatus
-from pre_award.assessment_store.db.models.comment.enums import CommentType
 from pre_award.assessment_store.db.models.flags.assessment_flag import AssessmentFlag
 from pre_award.assessment_store.db.models.flags.flag_update import FlagStatus, FlagUpdate
 from pre_award.assessment_store.db.queries.assessment_records._helpers import derive_application_values
 from pre_award.assessment_store.scripts.derive_assessment_values import derive_assessment_values
 from pre_award.config import Config
 from services.notify import NotificationService
-from tests.pre_award.application_store_tests.conftest import create_comment_with_updates
 from tests.pre_award.assessment_store_tests.test_assessment_mapping_fund_round import COF_FUND_ID
 from tests.pre_award.utils import AnyStringMatching
 
@@ -910,122 +904,3 @@ def test_derive_application_values_pfn(pfn_application_json_extract, app):
     assert result["funding_amount_requested"] == 12000 + 18000
     assert result["language"] == "en"
     assert result["project_name"] == "Test Council"
-
-
-def test_export_comments_to_excel(
-    db,
-    setup_submitted_application,
-    app,
-    mock_comments_sub_criteria,
-    comments_test_account,
-    comments_base_time,
-    comments_data,
-):
-    application_id = setup_submitted_application
-    user_id = comments_test_account.id
-
-    # Create comments with updates for the test application
-    for idx, cdata in enumerate(comments_data):
-        create_comment_with_updates(
-            db=db,
-            application_id=application_id,
-            user_id=user_id,
-            sub_criteria_id=cdata["sub_criteria_id"],
-            comment_type=cdata["comment_type"],
-            update_texts=cdata["updates"],
-            comments_base_time=comments_base_time + timedelta(hours=idx),
-            account=comments_test_account,
-        )
-
-    # Create a different application to later check for filtering
-    other_app = Applications(
-        id="00000000-0000-0000-0000-000000000000",
-        account_id="other-user",
-        fund_id=str(uuid4()),
-        round_id=str(uuid4()),
-        key="other-key",
-        language="en",
-        reference="OTHER-REF",
-        project_name="Other Project",
-    )
-    db.session.add(other_app)
-    db.session.commit()
-
-    # Create an assessment record for the other application
-    assessment_record = AssessmentRecord(
-        application_id=other_app.id,
-        short_id="OTHER-SHORT-ID",
-        type_of_application="Test",
-        project_name="Other Project",
-        funding_amount_requested=0,
-        round_id=other_app.round_id,
-        fund_id=other_app.fund_id,
-        language="en",
-        workflow_status="NOT_STARTED",
-        asset_type="Test",
-        jsonb_blob={},
-    )
-    db.session.add(assessment_record)
-    db.session.commit()
-
-    # Create a comment for the other application
-    create_comment_with_updates(
-        db=db,
-        application_id=other_app.id,
-        user_id="other-user",
-        sub_criteria_id=None,
-        comment_type=CommentType.WHOLE_APPLICATION,
-        update_texts=["Other app comment"],
-        comments_base_time=comments_base_time,
-        account=comments_test_account,
-    )
-    # Retrieve the test application
-    application = db.session.get(Applications, application_id)
-
-    # Retrieve comments for the test application
-    comments_list = retrieve_all_comments(
-        fund_id=application.fund_id,
-        round_id=application.round_id,
-        application_id=application.id,
-    )
-
-    # Export to Excel (in-memory)
-    fund_short_name = "TEST"
-    round_short_name = "ROUND"
-    with app.test_request_context():
-        response = export_comments_to_excel(
-            comments_list, fund_short_name, round_short_name, application_id=application_id
-        )
-        response.direct_passthrough = False
-        output = response.get_data()
-
-    # Read the Excel file
-    output_io = io.BytesIO(output)
-    df = pd.read_excel(output_io)
-
-    # Check that all comments for the test application are present
-    expected_comments = []
-    for cdata in comments_data:
-        expected_comments.extend(cdata["updates"])
-    assert Counter(df["Comment"]) == Counter(expected_comments)
-
-    # Check that no comments from the other applications are present
-    assert "Other app comment" not in df["Comment"].values
-
-    # Check ordering: Application-level comments (sub_criteria_id=None) first,
-    # then by sub_criteria_id, then by date_created
-    sub_criteria_col = df["Sub-criteria name"].fillna("None").values
-    app_level_end = 0
-    for val in sub_criteria_col:
-        if val == "None":
-            app_level_end += 1
-        else:
-            break
-
-    # All application-level comments should be at the top
-    assert all(val == "None" for val in sub_criteria_col[:app_level_end])
-    # The rest should be grouped by sub_criteria_id (SC1, then SC2)
-    sc1_start = app_level_end
-    sc1_end = sc1_start + sum(1 for c in comments_data if c["sub_criteria_id"] == "SC1" for _ in c["updates"])
-    assert all(val == "Sub Criteria 1" for val in sub_criteria_col[sc1_start:sc1_end])
-    assert all(val == "Sub Criteria 2" for val in sub_criteria_col[sc1_end:])
